main_classification/requirements.txt
pyserial==3.5
numpy==1.24.3
openpyxl==3.1.2
scipy==1.10.1
pyaudio>=0.2.11
librosa>=0.10.0
fastdtw>=0.3.4
pydub>=0.25.1
pandas>=1.5.0
torch>=1.13.0
transformers>=4.21.0
soundfile>=0.10.0